user-extraction-bot/requirement.txt
pyaes==1.6.1
pyrogram-dev==2.1.1
PySocks==1.7.1
TgCrypto==1.2.5
